pyproject.toml
[tool.poetry]
name = "clear-background-bot"
version = "0.1.0"
description = ""
authors = ["Kirill8769 <tkkrozn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.4"
onnxruntime-gpu = "^1.16.3"
rembg = {extras = ["gpu"], version = "^2.0.53"}
aiogram = "^3.3.0"
python-dotenv = "^1.0.0"
asyncpg = "^0.29.0"


[tool.poetry.group.lint.dependencies]
black = "^23.12.1"
flake8 = "^6.1.0"
isort = "^5.13.2"
mypy = "^1.8.0"

[tool.black]
line-length = 119

[tool.isort]
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = [".venv", "tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
